added an empty screen, user text select

+++ frontend/src/components/Details/Details.js
@@ -1,17 +1,25 @@
 import React, { Fragment } from 'react';
 import DateTime from './DateTime';
+import EmptyScreen from './EmptyScreen';
 import xss from 'xss';
 
-const Details = ({ showDays, currentDay, reloadDays, setScreen, deleteDay }) => {
+const Details = ({ showDays, currentDay, setCurrentDay, reloadDays, setScreen, deleteDay }) => {
   if (!currentDay) {
-    return <div></div>
+    return <EmptyScreen />
   }
 
   const removeDayBtn = () => {
+    if(!window.confirm('sure you wanna delete this?')) {
+      return
+    }
+    
     deleteDay(currentDay.key);
     reloadDays();
     setScreen('days');
+    setCurrentDay(null);
   }
+
+
   return (
     <div className="days__details">
       <div className="days__content">
